Add loading skeleton to job offers page
