+++ app/components/profile/offers/JobOffers.tsx
@@ -1,15 +1,12 @@
 "use client";
-
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { format } from "date-fns";
 import { toast } from "react-toastify";
 import { FaTrash, FaSave } from "react-icons/fa";
-
 interface Job {
   company: string;
   title: string;
 }
-
 interface JobOffer {
   job: Job;
   id: string;
@@ -17,49 +14,18 @@ interface JobOffer {
   offerDeadline: Date | null;
   salary: string;
 }
-
 interface JobOffersProps {
   jobOffers: JobOffer[];
   onEditOffer?: (offerId: string, updatedSalary: string) => void;
   onDeleteOffer?: (offerId: string) => void;
 }
-
 const Skeleton = ({ className }: { className: string }) => (
   <div
     className={`bg-zinc-800 motion-safe:animate-pulse rounded ${className}`}
   />
 );
-
-const SkeletonJobCard = () => (
-  <div className="bg-zinc-900 p-6 rounded-xl border border-zinc-800 shadow-sm">
-    <div className="space-y-4">
-      <div className="flex flex-col sm:flex-row justify-between items-start gap-4">
-        <div className="flex-shrink-0">
-          <Skeleton className="h-5 w-32 mb-2" />
-          <Skeleton className="h-5 w-48" />
-        </div>
-        <div className="flex-1">
-          <Skeleton className="h-4 w-40 mb-2" />
-          <Skeleton className="h-4 w-48 mb-2" />
-          <Skeleton className="h-8 w-full" />
-        </div>
-      </div>
-      <div className="flex justify-end space-x-3">
-        <Skeleton className="h-10 w-24" />
-        <Skeleton className="h-10 w-24" />
-      </div>
-    </div>
-  </div>
-);
-
 function JobOffers({ jobOffers, onEditOffer, onDeleteOffer }: JobOffersProps) {
   const [editingOffer, setEditingOffer] = useState<Record<string, string>>({});
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    setIsLoading(jobOffers.length === 0);
-  }, [jobOffers]);
-
   const formatSalaryDisplay = (value: string) => {
     const numericValue = value.replace(/\D/g, "");
     const formattedValue = numericValue.replace(/\B(?=(\d{3})+(?!\d))/g, ",");
@@ -69,21 +35,17 @@ function JobOffers({ jobOffers, onEditOffer, onDeleteOffer }: JobOffersProps) {
   const formatSalaryInput = (value: string) => {
     return value.replace(/[^\d.-]/g, "");
   };
-
   const handleOfferChange = (id: string, value: string) => {
     const formattedValue = formatSalaryInput(value);
     const displayValue = formatSalaryDisplay(formattedValue);
     setEditingOffer((prev) => ({ ...prev, [id]: displayValue }));
   };
-
   const getRawSalary = (formattedValue: string) => {
     return formattedValue.replace(/[^\d.-]/g, "");
   };
-
   const handleSaveOffer = async (id: string) => {
     const formattedSalary = editingOffer[id] || "";
     const rawSalary = getRawSalary(formattedSalary);
-
     if (rawSalary) {
       try {
         await onEditOffer?.(id, rawSalary);
@@ -98,13 +60,11 @@ function JobOffers({ jobOffers, onEditOffer, onDeleteOffer }: JobOffersProps) {
       }
     }
   };
-
   const handleDeleteOffer = async (offerId: string) => {
     const confirmed = window.confirm(
       "Are you sure you want to delete this offer?"
     );
     if (!confirmed) return;
-
     try {
       await onDeleteOffer?.(offerId);
       toast.success("Offer Deleted");
@@ -113,40 +73,6 @@ function JobOffers({ jobOffers, onEditOffer, onDeleteOffer }: JobOffersProps) {
       console.error("Error deleting offer:", error);
     }
   };
-
-  if (isLoading && jobOffers.length === 0) {
-    return (
-      <div className="w-full max-w-3xl mx-auto mt-6">
-        <div className="space-y-6 p-6">
-          {Array.from({ length: 3 }).map((_, index) => (
-            <SkeletonJobCard key={`loading-${index}`} />
-          ))}
-        </div>
-      </div>
-    );
-  }
-
-  if (jobOffers.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center py-16 text-gray-400">
-        <svg
-          className="w-12 h-12 mb-4"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
-          />
-        </svg>
-        <p>No Job Offers Found</p>
-      </div>
-    );
-  }
-
   return (
     <div className="w-full max-w-3xl mx-auto mt-6">
       <div className="space-y-6 p-6">
@@ -165,7 +91,6 @@ function JobOffers({ jobOffers, onEditOffer, onDeleteOffer }: JobOffersProps) {
     </div>
   );
 }
-
 interface JobOfferCardProps {
   offer: JobOffer;
   editingOffer?: string;
@@ -174,7 +99,6 @@ interface JobOfferCardProps {
   onDeleteOffer?: () => void;
   formatSalaryDisplay: (value: string) => string;
 }
-
 function JobOfferCard({
   offer,
   editingOffer,
@@ -249,5 +173,4 @@ function JobOfferCard({
     </div>
   );
 }
-
 export default JobOffers;
